Move ToolDefinition and related classes out of dartdoc_options
