+++ test/tool_runner_test.dart
@@ -8,6 +8,7 @@ import 'dart:io';
 
 import 'package:dartdoc/src/dartdoc_options.dart';
 import 'package:dartdoc/src/package_meta.dart';
+import 'package:dartdoc/src/tool_definition.dart';
 import 'package:dartdoc/src/tool_runner.dart';
 import 'package:path/path.dart' as path;
 import 'package:test/test.dart';
